+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.2.6] - 2019-02-13
 ### Added
+- `android_graphics.Bitmap.fromImage` static method
+- `android_graphics.Bitmap#getHeight()` method
+- `android_graphics.Bitmap#getWidth()` method
+- `android_graphics.Bitmap#toImage()` method
 
 ## [0.2.5] - 2019-02-13
 ### Added